python/pyproject.toml
[project]
name = "my-app"
version = "0.1.0"
description = "Internal CW application"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "pydantic==2.10.4",
    "structlog==24.4.0",
    "httpx==0.28.1",
    "python-jose[cryptography]==3.3.0",  # JWT validation for Okta OIDC
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "ruff==0.8.6",
    "bandit==1.8.3",
    "pip-audit==2.7.3",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "S",    # bandit (security)
    "B",    # bugbear
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "T20",  # print statements
]

[tool.pytest.ini_options]
testpaths = ["tests"]
